# openapi.json
{"openapi":"3.1.0","info":{"title":"CT AI Analysis API","version":"0.1.0"},"paths":{"/api/v1/users/login":{"post":{"tags":["users"],"summary":"Login","operationId":"login_api_v1_users_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_users_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/create_user":{"post":{"tags":["admin"],"summary":"Create User","operationId":"create_user_api_v1_admin_create_user_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminCreateUser"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/update_user/{user_id}":{"patch":{"tags":["admin"],"summary":"Update User","operationId":"update_user_api_v1_admin_update_user__user_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUpdateUser"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminUserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/admin/report-templates/upload":{"post":{"tags":["admin"],"summary":"Upload Report Template","operationId":"upload_report_template_api_v1_admin_report_templates_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_report_template_api_v1_admin_report_templates_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportTemplateOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/clinical-protocols/add":{"post":{"tags":["admin"],"summary":"Add Clinical Protocols","operationId":"add_clinical_protocols_api_v1_admin_clinical_protocols_add_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_clinical_protocols_api_v1_admin_clinical_protocols_add_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ClinicalProtocolOut"},"type":"array","title":"Response Add Clinical Protocols Api V1 Admin Clinical Protocols Add Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/reports/get_reports_by_login":{"get":{"tags":["reports"],"summary":"Get Id Reports","operationId":"get_id_reports_api_v1_reports_get_reports_by_login_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"login","in":"query","required":true,"schema":{"type":"string","title":"Login"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportsList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/my_reports":{"get":{"tags":["reports"],"summary":"Get My Reports","operationId":"get_my_reports_api_v1_reports_my_reports_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportsList"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/reports/{id_report}/add_review":{"post":{"tags":["reports"],"summary":"Add Review","operationId":"add_review_api_v1_reports__id_report__add_review_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id_report","in":"path","required":true,"schema":{"type":"string","title":"Id Report"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportReviewUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/{id_report}/view_html":{"get":{"tags":["reports"],"summary":"Get Html Report","operationId":"get_html_report_api_v1_reports__id_report__view_html_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id_report","in":"path","required":true,"schema":{"type":"string","title":"Id Report"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/{id_report}/pdf-url":{"get":{"tags":["reports"],"summary":"Get Pdf Report Url","operationId":"get_pdf_report_url_api_v1_reports__id_report__pdf_url_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id_report","in":"path","required":true,"schema":{"type":"string","title":"Id Report"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Pdf Report Url Api V1 Reports  Id Report  Pdf Url Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reports/{id_report}/status":{"get":{"tags":["reports"],"summary":"Get Report Status","operationId":"get_report_status_api_v1_reports__id_report__status_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"id_report","in":"path","required":true,"schema":{"type":"string","title":"Id Report"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Report Status Api V1 Reports  Id Report  Status Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/llm/create_report":{"post":{"tags":["llm"],"summary":"Create Report","operationId":"create_report_api_v1_llm_create_report_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_report_api_v1_llm_create_report_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AdminCreateUser":{"properties":{"login":{"type":"string","title":"Login"},"password":{"type":"string","title":"Password"},"role":{"$ref":"#/components/schemas/UserRole"},"organization_name":{"type":"string","title":"Organization Name"},"name":{"type":"string","title":"Name"},"surname":{"type":"string","title":"Surname"},"patronymic":{"type":"string","title":"Patronymic"},"date_of_birth":{"type":"string","format":"date","title":"Date Of Birth"}},"type":"object","required":["login","password","role","organization_name","name","surname","patronymic","date_of_birth"],"title":"AdminCreateUser"},"AdminUpdateUser":{"properties":{"role":{"anyOf":[{"$ref":"#/components/schemas/UserRole"},{"type":"null"}]},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"surname":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Surname"},"organization_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Organization Name"},"patronymic":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Patronymic"},"date_of_birth":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Date Of Birth"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"AdminUpdateUser"},"AdminUserOut":{"properties":{"id":{"type":"integer","title":"Id"},"login":{"type":"string","title":"Login"},"role":{"$ref":"#/components/schemas/UserRole"},"organization_name":{"type":"string","title":"Organization Name"},"name":{"type":"string","title":"Name"},"surname":{"type":"string","title":"Surname"},"patronymic":{"type":"string","title":"Patronymic"},"date_of_birth":{"type":"string","format":"date","title":"Date Of Birth"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","login","role","organization_name","name","surname","patronymic","date_of_birth","is_active"],"title":"AdminUserOut"},"Body_add_clinical_protocols_api_v1_admin_clinical_protocols_add_post":{"properties":{"files":{"items":{"type":"string","contentMediaType":"application/octet-stream"},"type":"array","title":"Files"}},"type":"object","required":["files"],"title":"Body_add_clinical_protocols_api_v1_admin_clinical_protocols_add_post"},"Body_create_report_api_v1_llm_create_report_post":{"properties":{"patient_name":{"type":"string","title":"Patient Name","description":"Patient full name"},"patient_sex":{"type":"string","title":"Patient Sex","description":"Sex (Male/Female)"},"birth_date":{"type":"string","format":"date","title":"Birth Date","description":"Date of birth"},"ct_date":{"type":"string","format":"date","title":"Ct Date","description":"CT study date"},"medical_text":{"type":"string","title":"Medical Text","description":"Medical history + symptoms","default":""},"enable_llm_judge":{"type":"boolean","title":"Enable Llm Judge","description":"Run LLM-as-judge after report generation","default":false},"ct_images":{"type":"string","contentMediaType":"application/octet-stream","title":"Ct Images","description":"ZIP archive with CT images"},"measurements_file":{"type":"string","contentMediaType":"application/octet-stream","title":"Measurements File","description":"Measurements file (CSV/JSON)"}},"type":"object","required":["patient_name","patient_sex","birth_date","ct_date","ct_images","measurements_file"],"title":"Body_create_report_api_v1_llm_create_report_post"},"Body_login_api_v1_users_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_users_login_post"},"Body_upload_report_template_api_v1_admin_report_templates_upload_post":{"properties":{"name":{"type":"string","title":"Name"},"version":{"type":"string","title":"Version"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"template_file":{"type":"string","contentMediaType":"application/octet-stream","title":"Template File"},"is_active":{"type":"boolean","title":"Is Active","default":false}},"type":"object","required":["name","version","template_file"],"title":"Body_upload_report_template_api_v1_admin_report_templates_upload_post"},"ClinicalProtocolOut":{"properties":{"title":{"type":"string","title":"Title"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"file_object_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"File Object Key"},"id":{"type":"integer","title":"Id"},"status":{"$ref":"#/components/schemas/ClinicalProtocolStatus"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"uploaded_at":{"type":"string","format":"date-time","title":"Uploaded At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"},"indexed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Indexed At"}},"type":"object","required":["title","id","status","uploaded_at"],"title":"ClinicalProtocolOut"},"ClinicalProtocolStatus":{"type":"string","enum":["uploaded","indexing","indexed","failed","archived"],"title":"ClinicalProtocolStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReportCreateResponse":{"properties":{"id_report":{"type":"string","title":"Id Report"},"status":{"$ref":"#/components/schemas/ReportStatus"}},"type":"object","required":["id_report","status"],"title":"ReportCreateResponse"},"ReportOut":{"properties":{"id_report":{"type":"string","title":"Id Report"},"user_id":{"type":"integer","title":"User Id"},"template_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Template Id"},"status":{"$ref":"#/components/schemas/ReportStatus"},"input_files":{"additionalProperties":true,"type":"object","title":"Input Files"},"measurements":{"additionalProperties":true,"type":"object","title":"Measurements"},"meta":{"additionalProperties":true,"type":"object","title":"Meta"},"llm_response":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Llm Response"},"html_object_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Html Object Key"},"pdf_object_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Object Key"},"review_score":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Review Score"},"review_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review Text"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id_report","user_id","status","input_files","measurements","meta","created_at"],"title":"ReportOut"},"ReportReviewUpdate":{"properties":{"review_score":{"type":"integer","title":"Review Score"},"review_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Review Text"}},"type":"object","required":["review_score"],"title":"ReportReviewUpdate"},"ReportStatus":{"type":"string","enum":["processing","completed","failed"],"title":"ReportStatus"},"ReportTemplateOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"version":{"type":"string","title":"Version"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"is_active":{"type":"boolean","title":"Is Active"},"created_by_user_id":{"type":"integer","title":"Created By User Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","name","version","is_active","created_by_user_id","created_at","updated_at"],"title":"ReportTemplateOut"},"ReportsList":{"properties":{"reports":{"items":{"$ref":"#/components/schemas/ReportOut"},"type":"array","title":"Reports"}},"type":"object","required":["reports"],"title":"ReportsList"},"TokenOut":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenOut"},"UserRole":{"type":"string","enum":["admin","user"],"title":"UserRole"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/users/login"}}}}}}